Move mock products to own file
